修改了数据库uid，gid，oid的类型，添加取消订单接口

+++ src/main/java/icu/callay/entity/ShoppingCart.java
@@ -2,6 +2,8 @@ package icu.callay.entity;
 
 import java.util.Date;
 
+import com.baomidou.mybatisplus.annotation.IdType;
+import com.baomidou.mybatisplus.annotation.TableId;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -25,10 +27,11 @@ public class ShoppingCart implements Serializable {
 
     private Long uid;
 
-    private Integer gid;
+    private Long gid;
 
     @JsonFormat(pattern = "yyyy-MM-dd", timezone = "GMT+8")
     private Date createTime;
+
 
 }
